Renamed Degeneration for easier use
